Added DIS support
Add support for importing .DIS shapes, courtesy of Jobo's DIS to OBJ script. Changed add-on name to Import-Starsiege-Tribes instead of Import-DTS. Removed facemaps from the DTS importer since Blender dropped support for it.
